+++ CHANGELOG.md
@@ -3562,3 +3562,4 @@
 - [2026-08-13 09:01] test(test): implement widget integration tests for test
 - [2026-08-13 09:28] feat(cache): add async cache stream provider
 - [2026-08-13 09:59] feat(ui): integrate custom telemetry logger in ui
+- [2026-08-13 10:09] style(ui): adjust design tokens, typography, and contrast in ui